base32: Started working on `base32`
At the moment, all it does is take blocks of characters and
or it with a variable of type long. This is so that the whole
block of characters can be represented as a single long. This
is helpful because we are dealing with 5-bit blocks later down
the line.